plain graphite series

+++ pkg/dry/math_test.go
@@ -9,7 +9,7 @@ import (
 func TestMax(t *testing.T) {
 	assert := assert.New(t)
 
-	assert.Equal(1, Max(1, -1))
-	assert.Equal(2, Max(1, 2))
-	assert.Equal(3, Max(3, 3))
+	assert.Equal(int64(1), Max(1, -1))
+	assert.Equal(int64(2), Max(1, 2))
+	assert.Equal(int64(3), Max(3, 3))
 }
